remove the autoconf macro ax_lib_oracle and replace with good old code
The ax_lib_oracle macro did not work so well. If it fails to find the oracle
libs or includes, it just omit building the oracle driver instead of failing.

Also the configure switches were very different from the other driver switches.
Now oracle is enabled using --with-oracle and setting CPPFLAGS and LDFLAGS
properly just like the other drivers. The only difference is, that it defaults
to --without-oracle while the others are compiled by default.