Add username and parameter to connect function
The url of the database connection is logged as a debug message. Since the url
may contain the password, the password may be printed out in clear text. This
is a security issue.
Now the username and password can be passed as separate entities and if need
inserted into the url using placeholders %u and %p. The password is never
logged.